Update common.xml
Proposal to add direct battery charge and energy consumption reporting in addition or instead of the current percentage of battery.  

The problem with the battery percentage is that the on-board system in most cases does not natively know the size of the energy storage and relies on the (possibly) configured battery capacity value.  In case of changing the battery on the field there may be different battery size or there may be a different type of power source on board. In any case, the percentage value is secondary information and dependent on a pre-configured capacity value, which may or may not be correct or relevant.  

If the on-board system has capability to report any consumption numbers, it almost certainly is measuring current and voltage samples.  These are primary information available to be reported and should be reported.

Possibly it could make sense to split message 147 to two separate messages: 1) multi-cell battery individual cell voltage reporting 2) total voltage & current readings and derivatives.