MAVGraph: Add option to choose plot bg from active flight mode
Simple little beautification to make it easier to visualise the
effect of changing flight modes.  Only an APM colour map is
implemented at this time, it uses the same colour pallet as
MAVProxy's mavflightview.

At the moment this cannot be specified at the same time as
--xaxis as for that, background vregions don't make much sense.
You would have to track a per-marker mode/colour instead, and
while that's possible, it seems overkill at this stage.