MAVUtil: Recognize Tri, Hexa, Octa and Tradheli as using the ACM Mode Map
Without this change, MAVProxy et al doesn't work for copter frame types
other than quad.  This manifests as an inability to change modes and an
exception whenever one types an unrecognized command (as MAVProxy tries
to map them to mode commands by default)

To have true coverage of the ACM codebase, MAV_TYPE_ROCKET would need to
be added to the list also as ACM uses this for FRAME_SINGLE (MAVLink not
providing an actual Single class). I don't want to be the one to extend
that hackery in to generic MAVLink code though!

Types that are copters but aren't implemented by ArduCopter (e.g.
COAXIAL) are similarly left out of this PR.