pymavlink: Just moved the check for end of file up a couple of lines
I have a log file which this was failing on as we were accessing the
elements variable even though it hadn't been initialised.  I just
moved the check up a couple of lines before accessing elements so it
will return if we are past the end of all the lines.