generator: added mavlink_frame_char() interface
this is like mavlink_parse_char() but allows for bad CRC. That makes
it possible for a C mavlink program to forward messages it wasn't
built with support for. In MAVLink 0.9 the CRC depended only on
content and not on the XML, which made forwarding of unknown messages
possible. In MAVLink 1.0 unknown messages can't be forwarded as they
fail the CRC check.

It is very useful to be able to forward unknown messages now that
routing of MAVLink packets is more common

mavlink_frame_char() will return:

 MAVLINK_FRAMING_INCOMPLETE = 0 when incomplete parsing
 MAVLINK_FRAMING_OK = 1 when parsed and good CRC
 MAVLINK_FRAMING_BAD_CRC = 2 when parsed and bad CRC