Clarify MAV_FRAME_BODY* enum values.
First off, MAV_FRAME_BODY* is not in NED coordinates, it should be in XYZ relative to the body. NED only makes sense for local.
Also clarified how the BODY and BODY_OFFSET frames are different. I don't think there's currently a usecase for the difference with the current MAV_CMD enum, but I clarified it none-the-less.
Additionally, support for BODY* frames is removed from POSITION_TARGET_LOCAL_NED, since those frames are not in the LOCAL_NED frame.