Fix the documented range for VFR_HUD heading
The documented range for VFR_HUD heading appears to not agree with the ground station applications I tried (qgroundcontrol and APMPlanner2). The fact that  this field is signed (int16_t vs uint16_t for throttle) implies this could have been intended, but not added to the help text. I think this should probably be updated here to reflect what is displayed in practice.