SET_ATTITUDE_TARGET partial quaternion writes
Update comment in SET_ATTITUDE_TARGET for partial quaternion writes:
Attitude quaternion (w, x, y, z order, zero-rotation is 1, 0, 0, 0). You can not set an attitude and a rate at the same time so the rate becomes a flag for setting a partial quaternion. For example if setting only attitude.roll the type_mask should be 0b01111110 but data is only available in the q.roll field and you ignore body_roll_rate field