common.xml: ignore FLAG_SAFETY_ARMED in DO_SET_MODE
This adds an additional comment to the flag MAV_MODE_FLAG_SAFETY_ARMED
which deprecates its use as a command. Instead the command
MAV_CMD_COMPONENT_ARM_DISARM should always be used to change the armed
state like it is already the case for most (if not all) ground sattions.

The reasoning behind the change is that it is:
- not intuitive to change the armed/disarmed state in a "mode",
- it is dangerous if forgotten because the vehicle can fall from the sky
  or equally bad suddenly arm,
- it can lead to race conditions because the flag needs to be set using
  the base_mode which arrives by the heartbeat (usually at only 1Hz), so
  by the time a DO_SET_MODE command is sent, the state might have
  changes on the vehicle.