Add MAVLink version handshake
When a radio or similar device is MAVLink 1 enabled but does not support MAVLink 2 and is connected between two nodes supporting MAVLink 2 it is nontrivial to discover if MAVLink 2 can be enabled: Both ends of the communication channel support it and show their support in the MAV_PROTOCOL_CAPABILITY flags but enabling it would still make the radio fail.
This extension allows to send a MAVLink 2 message as a test to see if it passes correctly through the link. As the request for the message is implemented as a command which should be NACKed as unsupported any node not having added support for this would default to MAVLink 1. This should allow a safe and robust handshake, including command retransmission if required.