camera: separate commands and settings
It turns out, it's a bad idea to have the resolution settings in the
same command as the start recording / capture command. This is because
both, the setting or the action, can fail and the result won't tell
which one did. Also applying a setting can take some time and can delay
the command to take a picture / record a video which should be
instantanteous.