Fix typo in explanation of vz
The description of GLOBAL_POSITION_INT mentions the following: "The position is in GPS-frame (right-handed, Z-up)". Now the explanation used for vz is saying "Ground Z Speed (Altitude, positive down), expressed as m/s * 100". One of the two is fault. The simulator is giving Z-up. I.e. the vz description is fault.