common: switch battery remaining time undefined to 0
Thanks to a review comment by @WickedShell I learned that extensions
only available in MAVLink 2 get filled with zeroes in older MAVLink
versions. Hence 0 has to be the value meaning no useful data is provided.
It's a small scrifice to have 1 second battery usage remaining defined
as the value for a completely empty battery.