Add non NED/ENU frames for external position
External pose as well as odometry information is almost never aligned with
earth's north and east directions. This adds a two new frames that are labeled with
FRD and FLU. With this we get rid of the ambigious usage of ENU and NED.
Additionally, this commit reduced the number of frames for external pose/odometry information.
The source for the information is now handled by an additional field in the odometry message.
The estimator type is extended to cover mocap and autopilot (fcu).