Extend ATTITUDE_QUATERNION with optional representation offset
The new field is a rotation offset by which the attitude quaternion and angular
speed vector should be rotated for user display (quaternion with [w, x, y, z]
order, zero-rotation is [1, 0, 0, 0], send [0, 0, 0, 0] if uneeded).
This field is intended for systems in which the reference attitude may change
during flight. For example, tailsitters VTOLs rotate their reference attitude
by 90 degrees between hover mode and fixed wing mode, thus repr_offset_q is
equal to [1, 0, 0, 0] in hover mode and equal to [0.7071, 0, 0.7071, 0] in
fixed wing mode.