common: add failure injection command with enums
In order to test failsafe mechanisms it would be nice to be able to
inject failures using a command.

Note that for safetey and security, autopilots should only accept such a
failure injection command if this has been specifically enabled using
e.g. a parameter or even a flag set at build time.

This commit is just a first suggestion how this could look like and in
no means exhaustive.