common: fix ambiguity in LOG_REQUEST_LIST
This specifies how to answer to LOG_REQUEST_LIST if there is no log
available. Without this response to signal that there is no log file the
requester has to assume there was a timeout and try again and again.

This is already the implementation in PX4 and from code inspection
ArduPilot seems to do the same.