common: add intent flags
@olliw42 discovered that the current gimbal v2 proposal still does not
properly address deconfliction of various sources. The rule was to
favor commands over the message streamed at a certain rate. This worked
for the case where manual control is overridden by a command as part of
a mission, however, it does not support e.g. a smart shot being sent
from a companion computer using the message (and not the command).

Therefore, the idea here is to add flags for the intent of the source.
This then lets the gimbal manager decide which intent it wants to allow
at any given time. The default order of the intents would be:
PILOT
GIMBAL_OPERATOR
SMARTSHOT
MISSION
TEST
CUSTOM1
CUSTOM2
with PILOT being the lowest and CUSTOM2 the highest "priority". However,
a gimbal manager is allowed to be configured differently depending on
the use case.

Clients sending gimbal control or nudging messages or commands are
supposed to set NUDGE or CONTROL if they want to start nudging or
controlling and follow up with at least one message with both NUDGE and
CONTROL unset in order to give up their control.

This proposal is by no means perfect but I argue it's better than the
current state and should enable many use cases, as well as make the
deconfliction rules easier to grasp.