common: fix PARAM_EXT trimming
It turns out that PARAM_EXT messages were not trimmed as we thought they
would be. The fields of a message are ordered by the size of the type,
however it does not matter if the type is an array or not.
This means that char[128] is not at the end of the message and
therefore not automatically trimmed.

This patch introduces new messages called _TRIMMED which has the
param_value as part of the extension meaning that it will be at the end
and trimmed correctly.