gimbal: add configure message
The deconfliction of various gimbal inputs has proven harder than
anticipated. Therefore, I suggest to go back to a configure message.
However, this new configure message only sets who is in control of a
gimbal but not what mode or how. The way how it is controlled is then up
to the components sending whatever messages and the gimbal manager
implementing the mixing logic for primary and secondary inputs.