gimbal: remove flags no longer needed
Since now we have a CONFIGURE command to set the components in control,
we no longer need the flags for nudge and override.

Also, we don't need the flag to set the focal length scaling anymore
because we can use the GIMBAL_MANAGER_SET_MANUAL_CONTROL message to
tilt/pan for this mode and it is up to the gimbal manager to scale it
according to mode/settings/focal length, depending on implementation.