Gimbal v2 changes (#1466)
* gimbal: add configure message

The deconfliction of various gimbal inputs has proven harder than
anticipated. Therefore, I suggest to go back to a configure message.
However, this new configure message only sets who is in control of a
gimbal but not what mode or how. The way how it is controlled is then up
to the components sending whatever messages and the gimbal manager
implementing the mixing logic for primary and secondary inputs.

* gimbal: fix outdated deprecation notice

* gimbal: add manual control message

We realized that it often doesn't make sense to send an angular rate or
angle from ground station or remote control but really it's just a
unitless input which then gets applied by the gimbal to whatever makes
sense. For instance, the gimbal might have speed or smoothness
parameters for this, or it might take into account the zoom level of a
camera to scale its angular rate. Therefore, we propose to have unitless
input for that case.

* gimbal: remove flags no longer needed

Since now we have a CONFIGURE command to set the components in control,
we no longer need the flags for nudge and override.

Also, we don't need the flag to set the focal length scaling anymore
because we can use the GIMBAL_MANAGER_SET_MANUAL_CONTROL message to
tilt/pan for this mode and it is up to the gimbal manager to scale it
according to mode/settings/focal length, depending on implementation.

* gimbal: specify angle limits as hardware limits

* gimbal: status of which components are in control

This allows a MAVLink component to know if it is currently in control or
not and wheterh it needs to send the GIMBAL_MANAGER_CONFIGURE command.

* gimbal: address review comments

* Minor subedit for consistency

Co-authored-by: Hamish Willee <hamishwillee@gmail.com>