gimbal: introduce special values for control
In order to make the control configuration more versatile, I added two
special values:

-1: leave unchanged:
This can be used to e.g. add sysid/compid to
secondary control without actually knowing the primary control and
preventing races.

-2: set itself in control:
This can be used where the configure command is used as part of a
mission and at the time where the mission is configured the sysid/compid
of the autopilot is not yet known, e.g. if the mission is created
offline.

-3: remove control if currently in control:
This allows to make sure something is not in control without keeping
track if it even was in control.