common: abolish PARAM_EXT_*_TRIMMED
It turns out that supporting the _TRIMMED message set for PARAM_EXT is
not as straightforward as it should be:
- Having to implement both variants of the messages ends up being a bit
  of code/binary bloat.
- When doing a param set without previously having done a get or request
  list is tricky because it is not clear whether TRIMMED can be used.
  This means that retransmission is required in case the wrong one is
  picked which makes it more brittle and can add some latency.

Furthermore, the changes have not actually been adopted anywhere that
I'm aware of, and they are not documented yet either.

I therefore came around and changed my opinion and vote to get rid of
these changes again.