common: request multiple CAMERA_IMAGE_CAPTURED
This adds a way to request multiple CAMERA_IMAGE_CAPTURED messages. This
solves the problem where it takes a long time to request all of the
messages if there are many pictures in storage.

For instance to request around 1000 pictures, it can easily take several
minutes depending on the link roundtrip time.