camera/component information: URI zero termination (#1807)
* camera_information: make URI zero terminated

This way it's much easier to deal with this message as the last char can
just be set to '\0' before handing it. Otherwise it needs to be copied
first.

* component_information: make URI zero terminated

This way it's much easier to deal with this message as the last char can
just be set to '\0' before handing it. Otherwise it needs to be copied
first.