common: remove option for relative speed change (#1812)
* common: remove option for relative speed change

This removes the functionality to do a speed change in a relative way
(e.g. 5 m/s slower) rather than an absolute way (10 m/s).

This option in param4 had been added in 2016 but neither implemented in
PX4 nor ArduPilot from what I can tell.

While this option might be a good idea, there is - at this point - no
safe option to add the functionality. If we were to start using the
option, then it might not work correctly if the autopilot does not
actually check param4.

Additionally to the fact that we can't start using it, the API is also
odd in that a negative speed (or at least -1) is generally ignored, so a
relative change that is slower (negative) would likely not work either.

I think the safe bet is to reserve this to 0, and add a new command for
relative changes, if required.

* Further modify MAV_CMD_DO_CHANGE_SPEED to use attribute to indicate reserved and default value

* Format using script

Co-authored-by: Hamish Willee <hamishwillee@gmail.com>