gimbal: more flexible attitude messages
This is an attempt to improve the attitude setpoint and attitude status
messages for gimbals, as suggested by olliw42.

The changes break the coupling between "locking" an angle so it stays
the same relative to Earth and the setpoint used and reported to do so.

Previously, these setpoints were possible:

- Pan: gimbal orientation relative to body (in body frame)
- Lock: gimbal orientation relative to earth (in earth frame)

The changes add one more setpoint:

- Lock: gimbal orientation relative to earth (in body frame)

Concerning the attitude status output from the gimbal, it means that the
a gimbal can provide the angle either body or earth frame (depending on
what it has available), and provide the yaw offset between the two, if
available.

This commit tries to add this without breaking existing implementations
by adding two flags, and an extension.
The alternative would be to change/replace the existing messages which
could lead to a difficult transition but would make the explanations and
descriptions easier.