MANUAL_CONTROL: Redefine the use of `z` (throttle) independent to thrust
Previously, the `z` component definition relied on having a knowledge of
which direction the thrust would result in, depending on it's value.

However this is a bad practice as the ManaulControl message itself
should be agnostic to a vehicle type (either reverse-thrustable ones
like Rover, or ones with only positive thrust like conventional
quadcopters).

With this new definition, the MANUAL_CONTROL message will not be
vehicle-specific and truly represent the Joystick / RC Transmitter's
throttle position