common.xml: exclude any stored home location from MISSION_CURRENT.total
ArduPilot stores its home location in slot-0.

Before this change, in the case you have that home location and a single waypoint uploaded ArduPilot would report a total of 2 - but when running the mission towards this single waypoint would report going to waypoint 1.  This is contrary to expectations of the user, so exclude slot 0 from the count.