cmake: generate headers at build time
This adds the generation of the MAVLink headers at build time.

Sadly it doesn't quite work the way I would want it to but instead it
doesn't generate anything on install unless I add ALL to the custom
target which means it's always being generated.

This has to do with the fact that it's an INTERFACE library.

Signed-off-by: Julian Oes <julian@oes.ch>