common.xml: deprecate the _INT frames (#2092)
* common.xml: deprecate the _INT frames

the frame of an altitude is not dependent on how the altitude is being transported (how other fields in the same message are being transported).

ArduPilot has treated these frames to be the same as the non-_INT equivalents in most places for a very long time.

We can save some confusion and perhaps a small amount of flash space if we remove the _INT frames

* Replace the deprecated types where they are used

* Rename lat/lon from X pos/Y pos in WGS84 frame

* Update message_definitions/v1.0/common.xml

* simply the global frames

* Apply suggestions from code review

* Apply suggestions from code review

* Update common.xml

---------

Co-authored-by: Hamish Willee <hamishwillee@gmail.com>