common.xml: remove recently added MAV_RESULT_PERMISSION_DENIED:
This was needed for the new in development multi GCS protocol.
It was decided to hold on this new result for the moment, in
order to make the current implementation easier to adopt, and
evaluate again in the future if this is needed.

For the situations meant to use this, we will use RESULT_FAILED