common: added options for hover vs loiter for DO_REPOSITION
this gives the GCS operator the choice of asking the aircraft to
either enter a hover or a fixed wing loiter at the target location.

In ArduPilot this is controlled at the moment using the Q_GUIDED_MODE
parameter, but it will be much better to be able to control this using
a UI element in the FlyTo dialog.